Cleaned up Redis query.
Was previously making an awful-looking ZRANGE query followed by an
additional query in the callback. A 3-second timeout was involved, if that's
any indication of how how fugly it was. Now it uses advice documented here:
http://stackoverflow.com/questions/8224548/cleaner-way-to-perform-several-redis-queries-within-node-js

This boils down to something along the lines of
SORT TIMESTAMPS GET # GET event:*->pev GET event:*->ev