Keeps track of duration of events.
Changed a number of things. Events should now contain attributes "start",
"name", and, "duration". Will no longer be identifying events with a
"name:subname" system, because it's dumb. All events will just follow the
convention of seperating hierarchical levels with colons. An event named
"reading:newspaper:nytimes" would be considered a "reading" event, a
"reading:newspaper" event, and a "reading:newspaper:nytimes" event.

The "duration" attribute of events is calculated at time of insertion. When
inserting a new event, we can look at the events sitting immediately before
and after it on the overall timeline of events. Using the start times of
these events we can update either the just-inserted event's duration, the
previous event's duration, or both.