Implemented public access tokens with expiration.
New form at /tokens allows user to enter in list of URLs (in the form of
relative paths) and an associated lifespan in seconds, minutes, hours, or
days. Can generate links of the form DOMAIN:PORT/some/path?token=12345...
where the "token" value is used in the checkAuth function to authorize
against. Links will only work within the given lifespan of the token.